Abstracted mutations into a seperate file.  Need clarification on what the 'spread operator is doing (passing parameters? to make it work??)  Works, but somewhat unclear.
